Fixed paths for easier user navigation in urls.py
