+++ NewsFeedApp/urls.py
@@ -14,13 +14,11 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
 from . import views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', views.home, name='home'),
-    path('login/', views.login, name='login'),
-    path('registration/', views.registration, name='registration'),
-    path('search/', views.query, name='query'),
+    path('user/', include('user.urls')),
 ]
