+++ NewsFeedApp/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
+from django.contrib.auth import authenticate, login as auth_login, logout as auth_logout
+from django.contrib import messages
 
 def home(request):
     if not request.user.is_authenticated:
@@ -7,11 +9,6 @@ def home(request):
 
     return redirect('/user/dashboard')
     
-def login(request):
-    return render(request, 'login.html', {'title':'Login'})
 
-def registration(request):
-    return render(request, 'registration.html', {'title':'Sign Up'})
 
-def query(request):
-    return render(request,'query.html', {'title':'Search'})
+
