+++ templates/home.html
@@ -496,7 +496,7 @@
 			</div>
 			<div class="logo">NewsFeed</div>
 			<a href="#"><i class="fa fa-search" id="search"></i></a>
-			<div class="log-in" id="log-in-btn">Log In</div>
+			<a href="login/"><button class="log-in" id="log-in-btn">Log In</button></a>
 			<div class="sign-up" id="sign-up-btn">Sign Up</div>
 			
 		</nav>
