+++ user/templates/login.html
@@ -1,0 +1,139 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <title>NewsFeed Login </title>
+    <meta name="description" content="">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;700&display=swap" rel="stylesheet">
+    <style>
+        .container {
+        font-family: Inter, Helvetica, sans-serif;
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        margin-right: -50%;
+        transform: translate(-50%, -50%);
+      
+        border-width: 1px;
+        border-radius: 8px;
+        border-style: solid;
+        border-color: black;
+      
+        height: 420px;
+        width: 800px;
+      }
+      
+      .logo {
+        font-weight: 700;
+      }
+      
+      .login-content {
+        margin-top: 25px;
+        display: flex;
+        flex-direction: column;
+        justify-content: space-evenly;
+        align-items: center;
+      }
+      
+      label {
+        text-align: left;
+        padding-top: 10px;
+        padding-bottom: 10px;
+      }
+      
+      input {
+        width: 693px;
+        height : 30px;
+        margin-top: 10px;
+        margin-bottom: 10px;
+      }
+      button {
+        color: white;
+        cursor: pointer;
+        display: flex;
+        flex-direction: row;
+        justify-content: center;
+        align-items: center;
+        padding: 0px 12px;
+        gap: 6px;
+      
+        width: 700px;
+        height: 40px;
+      
+        /* Primary */
+      
+        background: #074EE8;
+        border-radius: 4px;
+        border-color: #074EE8;
+        font-size: 15px;
+        border-style: solid;
+      
+        /* Inside auto layout */
+      
+        flex: none;
+        order: 4;
+        flex-grow: 0;
+      
+        margin-top: 10px;
+      }
+      
+      .text {
+        color: #074EE8;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        gap: 315px;
+      
+        width: 700px;
+        height: 45px;
+      
+        margin-top: 30px;
+        cursor: pointer;
+        /* Inside auto layout */
+      }
+      
+      .newuser:hover{
+        text-decoration: underline;
+      }
+      
+      .forgotpass:hover{
+        text-decoration: underline;
+      }
+      
+      .input-container {
+        margin-left: 50px;
+      }
+    </style>
+  </head>
+
+  <body>
+    <div class="container">
+      
+      <div class="login-content">
+        <div class="logo">
+          <h1>NewsFeed</h1>
+        </div>
+        <div class="input-container">
+          <label for="uname"><b>Username/Email</b></label>
+          <input type="text" name="uname" required>
+        </div>
+        <div class="input-container">
+          <label for="psw"><b>Password</b></label>
+          <input type="password" name="psw" required>
+        </div>
+        <div>
+          <button type="submit">Sign In</button>
+        </div>
+        <div class="text">
+          <a class="newuser" href="#">New user? Sign Up</a>
+          <a class="forgotpass">Forgot Password?</a>
+        </div>
+      </div> 
+    </div>
+  </body>
+</html>
